Alter format price, keys localStorage

+++ back-end/src/database/models/product.js
@@ -1,7 +1,7 @@
 const Product = (sequelize, DataTypes) => {
   const newProduct = sequelize.define('product', {
     name: DataTypes.STRING,
-    price: DataTypes.DOUBLE,
+    price: DataTypes.DECIMAL(10,2),
     url_image: DataTypes.STRING
   },
   { timestamps: false });
